refactor: move useAssetMetadata
